https://jira.catena-x.net/browse/KA-274 introduce alerts for warnings generated by the agent/skill framework.
