+++ infrastructure/consumer/resources/dataplane.properties
@@ -28,5 +28,5 @@ edc.vault.tenantid={{ .Values.security.vaultTenant }}
 
 edc.web.rest.cors.enabled=true
 edc.web.rest.cors.headers=origin, content-type, accept, authorization, x-api-key
-# edc.web.rest.cors.methods=
-# edc.web.rest.cors.origins=
+edc.web.rest.cors.methods=GET, PUT, POST, DELETE, OPTIONS
+edc.web.rest.cors.origins=*
