+++ infrastructure/docker-compose.yml
@@ -76,8 +76,12 @@ services:
       context: ../dataspace/agents/conforming
     image: ghcr.io/catenax-ng/product-knowledge/conforming-agent:0.8.1-SNAPSHOT
     platform: ${DOCKER_PLATFORM:-linux/amd64}
+    environment:
+      # Uncomment to enable for debugging, please choose a different port for each endpoint
+      JAVA_TOOL_OPTIONS: -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=*:4040
     ports:
       - "8898:8080" # sparql endpoint
+      - "4038:4040" # sparql endpoint
 
   #
   # Configures the provider agent
@@ -146,8 +150,8 @@ services:
       - "8700:8080" # AAS-API
       - "4048:4040" # DEBUG
     environment:
-      #PROVIDER_SPARQL_ENDPOINT: http://oem-data-plane:8082/api/agent
-      #PROVIDER_AGENT_PLANE: http://provider-agent:8082/sparql
+      PROVIDER_SPARQL_ENDPOINT: http://oem-data-plane:8082/api/agent
+      PROVIDER_AGENT_PLANE: http://provider-agent:8082/sparql
       JAVA_TOOL_OPTIONS: -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=*:4040
       TIMEOUT_SECONDS: 30
     #volumes:
