+++ infrastructure/oem/resources/controlplane2.properties
@@ -31,17 +31,17 @@ edc.ids.maintainer=http://catenax.net
 edc.ids.security.profile=base
 edc.ids.title=OEM Dataspace Connector
 
-edc.oauth.client.id={{ .Values.dataspace.dapsClientIdOem }}
-edc.oauth.private.key.alias={{ .Values.dataspace.privateKeyAliasOem }}
+edc.oauth.client.id=42:32:FD:1E:5F:35:9C:97:C9:E2:61:23:83:91:89:8D:7D:82:96:BD:keyid:42:32:FD:1E:5F:35:9C:97:C9:E2:61:23:83:91:89:8D:7D:82:96:BD
+edc.oauth.private.key.alias=oem-daps-key
 edc.oauth.provider.audience=idsc:IDS_CONNECTORS_ALL
 edc.oauth.endpoint.audience=http://oem-control-plane2:8282/api/v1/ids/data
 
 edc.data.encryption.algorithm=AES
 edc.data.encryption.keys.alias=oem-symmetric-key
 
-edc.oauth.provider.jwks.url={{ .Values.security.dapsUrl }}/.well-known/jwks.json
-edc.oauth.public.key.alias={{ .Values.dataspace.publicKeyAliasOem }}
-edc.oauth.token.url={{ .Values.security.dapsUrl }}/token
+edc.oauth.provider.jwks.url=https://daps1.int.demo.catena-x.net/.well-known/jwks.json
+edc.oauth.public.key.alias=oem-daps-crt
+edc.oauth.token.url=https://daps1.int.demo.catena-x.net/token
 
 edc.receiver.http.auth-code=X-Api-Key
 edc.receiver.http.auth-key=bar
@@ -50,7 +50,7 @@ edc.receiver.http.endpoint=http://oem-data-plane:8186/callback/endpoint-data-ref
 edc.dataplane.selector.agentplane.url=http://oem-data-plane:8082/
 edc.dataplane.selector.agentplane.sourcetypes=urn:cx:Protocol:w3c:Http#SPARQL,HttpData
 edc.dataplane.selector.agentplane.destinationtypes=HttpProxy,HttpProtocol
-edc.dataplane.selector.agentplane.properties={ "publicApiUrl": "http://oem-data-plane:8185/api/public" }
+edc.dataplane.selector.agentplane.properties={ "publicApiUrl": "http://oem-data-plane:8185/api/public/" }
 
 edc.transfer.proxy.endpoint=https://oem-data-plane:8082/
 edc.transfer.proxy.token.signer.privatekey.alias={{ .Values.dataspace.privateKeyAliasOem }}
@@ -63,6 +63,6 @@ edc.vault.tenantid={{ .Values.security.vaultTenant }}
 
 edc.web.rest.cors.enabled=true
 edc.web.rest.cors.headers=origin, content-type, accept, authorization, x-api-key
-# edc.web.rest.cors.methods=
-# edc.web.rest.cors.origins=
+edc.web.rest.cors.methods=GET, PUT, POST, DELETE, OPTIONS
+edc.web.rest.cors.origins=*
