+++ infrastructure/oem/resources/dataplane.properties
@@ -30,6 +30,6 @@ edc.vault.tenantid={{ .Values.security.vaultTenant }}
 
 edc.web.rest.cors.enabled=true
 edc.web.rest.cors.headers=origin, content-type, accept, authorization, x-api-key
-# edc.web.rest.cors.methods=
-# edc.web.rest.cors.origins=
+edc.web.rest.cors.methods=GET, PUT, POST, DELETE, OPTIONS
+edc.web.rest.cors.origins=*
