+++ infrastructure/resources/getReasonsAndEffectsOfDTC.rq
