+++ ux/skill_modules/src/components/EmptyResultBox/index.tsx
@@ -3,7 +3,8 @@ import { Typography } from 'cx-portal-shared-components';
 import React from 'react';
 import ErrorTwoToneIcon from '@mui/icons-material/ErrorTwoTone';
 
-export default function EmptyResultBox() {
+
+export const EmptyResultBox = () => {
   return (
     <Box textAlign="center" maxWidth="500px" ml="auto" mr="auto">
       <ErrorTwoToneIcon color="warning" fontSize="large" />
